Split out UCSF XML fetcher
